Added AddCarService and  JPA services

+++ src/main/java/com/example/securityumarket/services/page_service/StorageService.java
@@ -3,6 +3,8 @@ package com.example.securityumarket.services.page_service;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.*;
 import com.amazonaws.util.IOUtils;
+import com.example.securityumarket.exception.UAutoException;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -18,11 +20,18 @@ import java.util.Date;
 import java.util.Objects;
 
 @Slf4j
+@Getter
 @Service
 public class StorageService {
 
     @Value("${application.bucker.name}")
     private String bucketName;
+
+    @Value("${file.upload.minSize2}")
+    private long minFileSize;
+
+    @Value("${file.upload.maxSize}")
+    private long maxFileSize;
 
     private final AmazonS3 s3Client;
 
@@ -31,7 +40,7 @@ public class StorageService {
     }
 
     public byte[] downloadFile(String fileName) {
-        S3Object s3Object = s3Client.getObject(bucketName, fileName);
+        S3Object s3Object = s3Client.getObject(getBucketName(), fileName);
         S3ObjectInputStream inputStream = s3Object.getObjectContent();
         try {
             byte[] byteArray = IOUtils.toByteArray(inputStream);
@@ -43,7 +52,7 @@ public class StorageService {
     }
 
     public String deleteFile(String fileName) {
-        s3Client.deleteObject(bucketName,fileName);
+        s3Client.deleteObject(getBucketName(), fileName);
         return fileName + " removed.";
     }
 
@@ -58,38 +67,60 @@ public class StorageService {
     }
 
     public String uploadFileWithPublicRead(MultipartFile file) {
-        File fileObj = convertMultiPartFileToFile(file);
-        String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
-        s3Client.putObject(new PutObjectRequest(bucketName,fileName,fileObj)
-                .withCannedAcl(CannedAccessControlList.PublicRead));
-        fileObj.delete();
-        return "File uploaded : " + fileName;
+        if (file.getSize() >= minFileSize && file.getSize() <= maxFileSize) {
+            String contentType = file.getContentType();
+            if (Objects.requireNonNull(contentType).startsWith("image/") || contentType.startsWith("video/")) {
+                File fileObj = convertMultiPartFileToFile(file);
+                String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
+                s3Client.putObject(new PutObjectRequest(getBucketName(), fileName, fileObj)
+                        .withCannedAcl(CannedAccessControlList.PublicRead));
+                fileObj.delete();
+                return "File uploaded : " + fileName;
+            } else {
+                throw new UAutoException("Unsupported file type. Please upload photos or videos.");
+            }
+        } else {
+            throw new UAutoException("File size must be between " + minFileSize + " and " + maxFileSize + " bytes.");
+        }
     }
 
-    public String getPhotoUrlFromPublicRead(String fileName)
-    {
-        GeneratePresignedUrlRequest urlRequest = new GeneratePresignedUrlRequest(bucketName, fileName);
-        ResponseHeaderOverrides headers = new ResponseHeaderOverrides();
-        urlRequest.addRequestParameter("Content-Disposition", "inline");
-        URL url = s3Client.generatePresignedUrl(urlRequest);
-        return url.toString();
+    public String getFileUrlFromPublicRead2(String fileName) {
+        if (s3Client.doesObjectExist(getBucketName(), fileName)) {
+            GeneratePresignedUrlRequest urlRequest = new GeneratePresignedUrlRequest(getBucketName(), fileName);
+            ResponseHeaderOverrides headers = new ResponseHeaderOverrides();
+            urlRequest.addRequestParameter("Content-Disposition", "inline");
+            URL url = s3Client.generatePresignedUrl(urlRequest);
+            return url.toString();
+        } else {
+            throw new UAutoException("File with the name " + fileName + " does not exist.");
+        }
     }
+
+    public String getFileUrlFromPublicRead(String fileName) {
+        if (s3Client.doesObjectExist(getBucketName(), fileName)) {
+            GeneratePresignedUrlRequest urlRequest = new GeneratePresignedUrlRequest(getBucketName(), fileName);
+            URL url = s3Client.generatePresignedUrl(urlRequest);
+            return url.toString();
+        } else {
+            throw new UAutoException("File with the name " + fileName + " does not exist.");
+        }
+    }
+
 
     public String uploadFileWithPrivateRead(MultipartFile file) {
         File fileObj = convertMultiPartFileToFile(file);
         String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
-        s3Client.putObject(new PutObjectRequest(bucketName,fileName,fileObj));
+        s3Client.putObject(new PutObjectRequest(getBucketName(), fileName, fileObj));
         fileObj.delete();
         return "File uploaded : " + fileName;
     }
 
-    public String getPhotoUrlFromPrivateRead(String fileName)
+    public String getFileUrlFromPrivateRead(String fileName)
     {
-        GeneratePresignedUrlRequest urlRequest = new GeneratePresignedUrlRequest(bucketName, fileName);
+        GeneratePresignedUrlRequest urlRequest = new GeneratePresignedUrlRequest(getBucketName(), fileName);
         urlRequest.setExpiration(Date.from(Instant.now().plus(Duration.ofHours(1))));
         return s3Client.generatePresignedUrl(urlRequest).toString();
     }
-
 
 
 }
